tovabbi vegpontok definialasa

+++ src/main/java/com/example/BoardGameHub/controller/UserController.java
@@ -8,6 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -33,7 +34,7 @@ public class UserController {
     @GetMapping("/{userId}")
     public ResponseEntity<User> getUser(
         @PathVariable
-        Integer userId
+        Long userId
     ) {
         final Optional<User> userOptional = userService.findById(userId);
