+++ src/main/java/com/example/BoardGameHub/repository/BoardGameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardGameRepository extends JpaRepository<BoardGame, Integer> {
+public interface BoardGameRepository extends JpaRepository<BoardGame, Long> {
 }
